+++ tests/core/engines/alpha/steps/tools.py
@@ -338,6 +338,31 @@ def given_a_tool(
             "parameters": {},
             "required": [],
         },
+        "consult_policy": {
+            "name": "consult_policy",
+            "description": "",
+            "module_path": "tests.tool_utilities",
+            "parameters": {},
+            "required": [],
+        },
+        "other_inquiries": {
+            "name": "other_inquiries",
+            "description": "This tool needs to be run when looking for answers that are not covered by other resources",
+            "module_path": "tests.tool_utilities",
+            "parameters": {},
+            "required": [],
+        },
+        "try_unlock_card": {
+            "name": "try_unlock_card",
+            "description": "This tool unlocks a credit card",
+            "module_path": "tests.tool_utilities",
+            "parameters": {
+                "last_6_digits": {
+                    "type": "string",
+                },
+            },
+            "required": [],
+        },
     }
 
     tool = context.sync_await(create_tool(**tools[tool_name]))