+++ tests/tool_utilities.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 import enum
+from typing import Optional
 
 from parlant.core.tools import ToolResult
 
@@ -145,3 +146,24 @@ def fetch_account_balance() -> ToolResult:
 
 def get_keyleth_stamina() -> ToolResult:
     return ToolResult(data=100.0)
+
+
+def consult_policy() -> ToolResult:
+    policies = {
+        "return_policy": "The return policy allows returns within 4 days and 4 hours from the time of purchase.",
+        "warranty_policy": "All products come with a 1-year warranty.",
+    }
+    return ToolResult(policies)
+
+
+def other_inquiries() -> ToolResult:
+    return ToolResult("Sorry, we could not find a specific answer to your query.")
+
+
+def try_unlock_card(last_6_digits: Optional[str] = None) -> ToolResult:
+    try:
+        if not last_6_digits:
+            return ToolResult({"failure": "need to specify the last 6 digits of the card"})
+        return ToolResult({"success": "card succesfuly unlocked"})
+    except BaseException:
+        return ToolResult({"failure": "system error"})